+++ flutter_web_auth_2/test/flutter_web_auth_2_test.dart
@@ -4,8 +4,10 @@ import 'package:flutter_web_auth_2/flutter_web_auth_2.dart';
 
 void main() {
   const channel = MethodChannel('flutter_web_auth_2');
+  var expectedPreferAuthTabs = true;
 
   setUp(() {
+    expectedPreferAuthTabs = true;
     TestWidgetsFlutterBinding.ensureInitialized();
 
     TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
@@ -17,6 +19,10 @@ void main() {
         'https://example.com/login',
       );
       expect(methodCall.arguments['callbackUrlScheme'] as String, 'foobar');
+      expect(
+        methodCall.arguments['options']['preferAuthTabs'] as bool,
+        expectedPreferAuthTabs,
+      );
 
       return 'https://example.com/success';
     });
@@ -32,6 +38,19 @@ void main() {
       await FlutterWebAuth2.authenticate(
         url: 'https://example.com/login',
         callbackUrlScheme: 'foobar',
+      ),
+      'https://example.com/success',
+    );
+  });
+
+  test('can force a regular Custom Tab on Android', () async {
+    expectedPreferAuthTabs = false;
+
+    expect(
+      await FlutterWebAuth2.authenticate(
+        url: 'https://example.com/login',
+        callbackUrlScheme: 'foobar',
+        options: const FlutterWebAuth2Options(preferAuthTabs: false),
       ),
       'https://example.com/success',
     );